FIX: new numpy does raises exception on bincount of float arrays

+++ nilearn/_utils/ndimage.py
@@ -32,7 +32,7 @@ def largest_connected_component(volume):
         raise ValueError('No non-zero values: no connected components')
     if label_nb == 1:
         return volume.astype(np.bool)
-    label_count = np.bincount(labels.ravel())
+    label_count = np.bincount(labels.ravel().astype(np.int))
     # discard the 0 label
     label_count[0] = 0
     return labels == label_count.argmax()
